Got a rudimentary version of the projection matrix working.
There's a slight problem with this version, and it is an interesting one:
while it is acceptable for screen space's aspect ratio to be important (i.e.,
if your camera's field of view is not square, you ought to expect compression
in the minor axis), that's not what I had intended for viewport space.

Just because your viewport is rectangular does not mean that your shapes
should be squashed!  Use screen space for that!

This means that I'll need to preserve the screen space aspect ratio as I space
it up to view space.  Probably easy to do.  I'll give it a think later.