Got mouse aiming to work.
The ultimate soluution was to give up on the idea of maintaining a separate
camera basis that would only be adjusted in orientation by the mouse (i.e.,
the mouse representing relative pitch and yaw on the camera frame.)

Now, the mouse represents pitch and yaw with respect to the *absolute* frame
of reference -- the absolute X axis and the absolute Y axis.  T?hen we
construct a basis anew each from and aim it _once_ using the
eulerRotationMatrix().  Since the pitch and yaw area already used to aim the
camera, we can also tie changes in yaw to the A and D keys.

The only problem with this approach is that aiming is absolutely unlimited --
i.e., I can now face down and then move backward, and that is the same as
jumping in the air!  Movement forward and back ward needs to be tied to the
grid's frame of reference, so there's some more math to do.