Inspired by a discussion with the students yesterday, simplify the "allow boosting" logic.
A ∙ B == |A| |B| cos θ.  In this particular case, |B| is 1, and we
normalize |A| to 1 prior to calculating the dot product.  But _why_?  All
we're testing for in the end is that the vectors are "roughly pointing in the
same direction" (i.e., that the dot product is >= 0), and |A| doesn't matter
for that purpose.

So just use the dot product directly and the magnitude safety check goes away
entirely.