Completed the clip function for polygons.
Included the 'if' statements for the four ways that p and q could be oriented in relation to clipPlane.

Started to write a function for printing the max amount of characters for the vertices of clippedPolygon so we can see if the clipping function works.
TODO:
Finish and test the printing function.